Add Search module to log client searches
This version will keep log of the search term, a timestamp of the API
call, and the total results found from the API response meta data.

A function to get trending searches is also added, where the X most searched
terms should be displayed as a list. X is initially set to 100.